server/pyproject.toml
[tool.black]
line-length = 100
skip-string-normalization = true
target-version = ["py38"]
exclude = '\.eggs|\.git|\.mypy_cache|\.tox|\.env|\.venv|env|venv|_build|buck-out|build|dist'

[tool.isort]
profile = "black"
line_length = 100
# Sort by name, don't cluster "from" vs "import"
force_sort_within_sections = true
# Combines "as" imports on the same line
combine_as_imports = true

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
exclude = '\.eggs|\.git|\.mypy_cache|\.tox|\.env|\.venv|env|venv|_build|buck-out|build|dist'

[tool.poetry]
name = "UMD_server"
version = "1.0.0"
description = "UMD Data Server"
authors = ["Kitware, Inc. <viame-web@kitware.com>"]
license = "Apache 2.0"
include = ["**/*.mako"]
readme = "README.md"
keywords = ["UMD"]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]
packages = [
  { include = "UMD_server" },
  { include = "UMD_tasks" },
  { include = "../../dive/server/dive_server" },
  { include = "../../dive/server/dive_tasks" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
cheroot = ">=8.4.5"
click = "^8.1.3"
girder = "3.1.14"
girder_jobs = "3.1.14"
girder_worker = "0.8.1"
girder_worker_utils = "0.8.7"
pydantic = "1.9.0"
pyrabbit2 = "1.0.7"
typing-extensions = "^4.2.0"
GPUtil = "^1.4.0"
requests = ">=2.27.1"
urllib3 = "<1.27"
CherryPy = "^18.8.0"
Pillow = "^9.2.0"
image = "^1.5.33"
[tool.poetry.dev-dependencies]
numpy = "^1.21.4"
opencv-python = "^4.5.5"
tox = "^3.25.0"

[tool.poetry.plugins."girder.plugin"]
UMD_plugin = "UMD_server:UMDPlugin"

[tool.poetry.plugins."girder_worker_plugins"]
UMD_tasks = "UMD_tasks:UMDTasksPlugin"


[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
